Task: Refactor utils out of ui package
