+++ video/src/main/java/com/gilbertohdz/media3/video/utils/DummyPlayer.kt
@@ -1,4 +1,4 @@
-package com.gilbertohdz.media3.video.ui.utils
+package com.gilbertohdz.media3.video.utils
 
 import android.os.Looper
 import android.view.Surface
